requirements.txt
numpy==1.24.3
scikit-image
dominate>=2.8.0
Pillow>=10.0.0
wandb>=0.16.0
visdom>=0.1.8.8